Added Fireplace Story
Added Fireplace Story for anyone who is wanting to contribute to the project I hope I can finish this project sometime soon because this has been going on for way too long and I need to get a start on it. Thank you to whoever is going to contribute to this story.